frontend: rename mnemonicpassphrase.tsx to passphrase.tsx
