+++ frontends/web/src/routes/device/bitbox02/passphrase.tsx
